FakeNewsNet-master-fixed/fakenewsnet_fixed_hold/requirements.txt
beautifulsoup4>=4.11.0
certifi>=2022.12.7
chardet>=4.0.0
click>=8.0.0
cssselect>=1.2.0
feedfinder2>=0.0.4
feedparser>=6.0.8
Flask>=2.2.0
Flask-Cors>=3.0.10
idna>=3.4
Jinja2>=3.1.2
lxml>=4.9.0
MarkupSafe>=2.1.1
newspaper3k>=0.2.8
nltk>=3.7
oauthlib>=3.2.2
Pillow>=9.0.0
python-dateutil>=2.8.2
PyYAML>=6.0
requests>=2.28.0
requests-file>=1.5.1
requests-oauthlib>=1.3.1
six>=1.16.0
soupsieve>=2.3.2
tinysegmenter>=0.3
tldextract>=3.4.0
tqdm>=4.64.0
twython>=3.9.1
urllib3>=1.26.14
Werkzeug>=2.2.0
